feat(help): move help out of experimental
