+++ angular.json
@@ -479,6 +479,37 @@
         }
       }
     },
+    "covalent-help": {
+      "root": "src/platform/help",
+      "sourceRoot": "src/platform/help",
+      "projectType": "library",
+      "prefix": "lib",
+      "architect": {
+        "build": {
+          "builder": "@angular-devkit/build-ng-packagr:build",
+          "options": {
+            "tsConfig": "src/platform/help/tsconfig.lib.json",
+            "project": "src/platform/help/ng-package.js"
+          }
+        },
+        "test": {
+          "builder": "@angular-devkit/build-angular:karma",
+          "options": {
+            "main": "src/platform/help/test.ts",
+            "tsConfig": "src/platform/help/tsconfig.spec.json",
+            "karmaConfig": "src/platform/help/karma.conf.js",
+            "polyfills": "src/polyfills.ts"
+          }
+        },
+        "lint": {
+          "builder": "@angular-devkit/build-angular:tslint",
+          "options": {
+            "tsConfig": ["src/platform/help/tsconfig.lib.json", "src/platform/help/tsconfig.spec.json"],
+            "exclude": ["**/node_modules/**"]
+          }
+        }
+      }
+    },
     "covalent-experimental": {
       "root": "src/platform/experimental",
       "sourceRoot": "src/platform/experimental",
