+++ src/app/components/components/components.module.ts
@@ -102,7 +102,7 @@ import { ToolbarModule } from '../../components/toolbar/toolbar.module';
 
 import { TdTestDynamicComponent } from './dynamic-forms/dynamic-forms.component';
 import { HelpDemoComponent } from './help/help.component';
-import { CovalentHelpModule } from '@covalent/experimental/help';
+import { CovalentHelpModule } from '@covalent/help';
 import { CovalentSidesheetModule } from '@covalent/core/sidesheet';
 import { DraggableDemoComponent } from './dialogs/draggable-demo/draggable-demo.component';
