+++ src/app/components/components/help/help.component.html
@@ -43,7 +43,7 @@
           import {
             IHelpMenuDataItem,
             DraggableHelpWindowDialogService
-          } from '@covalent/experimental/help';
+          } from '@covalent/help';
 
           oneItem: IHelpMenuDataItem[] = [
             {
@@ -165,7 +165,7 @@
             IHelpMenuDataItem,
             DraggableHelpWindowDialogService,
             HelpWindowComponent,
-          } from '@covalent/experimental/help';
+          } from '@covalent/help';
 
           windowOpen: boolean = false;
           ref: MatDialogRef<HelpWindowComponent>;
@@ -200,4 +200,4 @@
   </mat-tab-group>
 </mat-card>
 
-<td-readme-loader resourceUrl="platform/experimental/help/README.md"></td-readme-loader>
+<td-readme-loader resourceUrl="platform/help/README.md"></td-readme-loader>
