+++ src/app/components/components/help/help.component.ts
@@ -1,7 +1,7 @@
 import { Component, HostBinding } from '@angular/core';
 
 import { slideInDownAnimation } from '../../../app.animations';
-import { IHelpMenuDataItem, DraggableHelpWindowDialogService, HelpWindowComponent } from '@covalent/experimental/help';
+import { IHelpMenuDataItem, DraggableHelpWindowDialogService, HelpWindowComponent } from '@covalent/help';
 import { MatDialogRef } from '@angular/material/dialog';
 
 @Component({
