--- src/platform/experimental/help/package.json
@@ -1,7 +1,0 @@
-{
-  "ngPackage": {
-    "lib": {
-      "entryFile": "index.ts"
-    }
-  }
-}
