+++ src/platform/help/draggable-help-window-dialog/draggable-help-window-dialog.service.ts
