+++ src/platform/help/help-window/help-window.component.ts
