+++ src/platform/help/help.component.scss
