+++ src/platform/help/help.component.spec.ts
