+++ src/platform/help/ng-package.js
@@ -1,0 +1,5 @@
+const ngPackageSettings = require('../ng-package-common.js');
+module.exports = {
+  ...ngPackageSettings,
+  dest: '../../../deploy/platform/help',
+};
