+++ src/platform/help/package.json
@@ -1,0 +1,35 @@
+{
+  "name": "@covalent/help",
+  "version": "0.0.0-COVALENT",
+  "description": "Teradata UI Platform Help Module",
+  "keywords": [
+    "angular",
+    "components",
+    "covalent",
+    "help",
+    "documentation",
+    "markdown",
+    "inline docs",
+    "docs"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/teradata/covalent.git"
+  },
+  "bugs": {
+    "url": "https://github.com/teradata/covalent/issues"
+  },
+  "license": "MIT",
+  "author": "Teradata UX",
+  "contributors": [
+    "Christian Memije <christian.memije@teradata.com>"
+  ],
+  "peerDependencies": {
+    "@angular/core": "^0.0.0-NG",
+    "@angular/common": "^0.0.0-NG",
+    "@angular/material": "^0.0.0-MATERIAL",
+    "@angular/cdk": "^0.0.0-MATERIAL",
+    "@covalent/core": "^0.0.0-COVALENT",
+    "@covalent/flavored-markdown": "^0.0.0-COVALENT"
+  }
+}
