+++ src/platform/help/public-api.ts
