+++ src/platform/help/test.ts
@@ -1,0 +1,6 @@
+// This file is required by karma.conf.js
+import { setupTestEnv } from '../../setupTestEnv';
+
+declare const require: any;
+
+setupTestEnv(require.context('./', true, /\.spec\.ts$/));
