+++ src/platform/help/tsconfig.lib.json
@@ -1,0 +1,3 @@
+{
+  "extends": "../tsconfig.lib.json"
+}
