+++ src/platform/help/tsconfig.spec.json
@@ -1,0 +1,5 @@
+{
+  "extends": "../../tsconfig.spec.json",
+  "files": ["test.ts", "../../polyfills.ts"],
+  "include": ["**/*.spec.ts", "**/*.d.ts"]
+}
